Wrap1P(): Change the return type to size_t.
DWORD is still a 32-bit type in 64-bit builds, and we certainly don't want to
truncate any potential 64-bit value returned from the wrapped function.

+++ src/wrappers.c
@@ -6,7 +6,7 @@
   * Generic call wrappers to cut down redundancy.
   */
 
-DWORD WINAPI Wrap1P(
+size_t WINAPI Wrap1P(
 	Wrap1PFunc_t func,
 	LPCSTR lpsz
 )
